+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to [bpmn-js-element-templates](https://github.com/bpmn-io/bp
 ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: support job priority via `zeebe:jobPriorityDefinition` binding
+* `FIX`: assign `$parent` property in all cases for I/O mapping ([#265](https://github.com/bpmn-io/bpmn-js-element-templates/pull/265))
 
 ## 2.27.0